+++ docs/md/manual/releasenotes.md
@@ -7,6 +7,7 @@
 ### Bug fixes
 
 - Prevent the generation of Windows reserved names. (`T3937`)
+- Prevent "Expecting type and name" parse error when using the `androidx.window` library in an Android project. (`T13715`)
 
 ## Version 7.2